Changed data type used to store line numbers from unsigned long to unsigned.
git-svn-id: svn://svn.cc65.org/cc65/trunk@5222 b7a2c559-68d2-44c3-8de9-860c34a00d81

+++ src/ld65/fragment.h
@@ -98,9 +98,9 @@ INLINE const char* GetFragmentSourceName (const Fragment* F)
 #endif
 
 #if defined(HAVE_INLINE)
-INLINE unsigned long GetFragmentSourceLine (const Fragment* F)
+INLINE unsigned GetFragmentSourceLine (const Fragment* F)
 /* Return the source file line for this fragment */
-{
+{               
     return GetSourceLineFromList (&F->LineInfos);
 }
 #else
